- second round of placemap and model classes fixes, i found out that the delete function cant be static.
git-svn-id: http://repos.svn.fusionticket.org/fusionticket/trunk@776 87ea6b0e-88cb-4cdd-95e9-9c88242927a5

+++ includes/admin/OrtView.php
@@ -41,7 +41,10 @@ class OrtView extends AdminView {
     $res = ShopDB::query($query);
     $alt = 0;
     echo "<table class='admin_list' width='$this->width' cellpadding='4' cellspacing='1'>\n";
-    echo "<tr><td class='admin_list_title' colspan='5' align='left'>" . con('ort_title') . "</td></tr>\n";
+    echo "<tr><td class='admin_list_title' colspan='2' align='left'>" . con('ort_title') . "</td>";
+    echo "<td colspan=1 align=right><a class='link' href='{$_SERVER['PHP_SELF']}?action=add'>
+            <img src='images/add.png' border='0' alt='".con('add')."' title='".con('add')."'></a></td>";
+    echo "</tr>\n";
     if (!$res) {
       //
     } else {
@@ -49,7 +52,7 @@ class OrtView extends AdminView {
           echo "<tr class='admin_list_row_$alt'>";
           echo "<td class='admin_list_item' width='50%'>{$row['ort_name']}</td>\n";
           echo "<td class='admin_list_item'>{$row['ort_city']}</td>\n";
-          echo "<td class='admin_list_item'width='40' nowrap><nowrap><a class='link' href='{$_SERVER['PHP_SELF']}?action=edit&ort_id={$row['ort_id']}'><img src='images/edit.gif' border='0' alt='" . edit . "' title='" . edit . "'></a>";
+          echo "<td class='admin_list_item'width='40' align='right' nowrap><nowrap><a class='link' href='{$_SERVER['PHP_SELF']}?action=edit&ort_id={$row['ort_id']}'><img src='images/edit.gif' border='0' alt='" . edit . "' title='" . edit . "'></a>";
           echo "<a class='link' href='javascript:if(confirm(\"" . con('delete_item') . "\")){location.href=\"{$_SERVER['PHP_SELF']}?action=remove&ort_id={$row['ort_id']}\";}'>
                  <img src='images/trash.png' border='0' alt='" . con('remove') . "' title='" . con('remove') . "'></a></nowrap></td>\n";
           echo "</tr>";
@@ -57,7 +60,6 @@ class OrtView extends AdminView {
       }
       echo "</table>\n";
     }
-    echo "<br><center><a class='link' href='{$_SERVER['PHP_SELF']}?action=add'>" . con('add') . "</a></center>";
     // echo "<br><center><a class='link' href='{$_SERVER['PHP_SELF']}?action=import'>".import."</a></center>";
   }
 
@@ -95,6 +97,7 @@ class OrtView extends AdminView {
   }
 
   function draw () {
+    addNotice("Action: {$_REQUEST['action']} pm_id {$_REQUEST['pm_id']}");
     if (preg_match('/_pm$/', $_REQUEST['action']) or preg_match('/_pmz$/', $_REQUEST['action']) or
         preg_match('/_pmp$/', $_REQUEST['action']) or preg_match('/_category$/', $_REQUEST['action'])){
         require_once('admin/PlaceMapView2.php');
@@ -116,8 +119,8 @@ class OrtView extends AdminView {
     } elseif ($_GET['action'] == 'add') {
        $this->form($row, $err, con('ort_add_title'));
     } elseif ($_GET['action'] == 'remove' and $_GET['ort_id'] > 0) {
-      $ort = ort::load($_GET['ort_id']);
-      $ort->delete();
+      if($ort = ort::load($_GET['ort_id']))
+        $ort->delete();
       $this->table();
     } elseif ($_GET['action'] == 'import') {
         $this->shared_ort_list();
